+++ changelog.md
@@ -1,5 +1,7 @@
 
-[v0.56.8](https://github.com/nextcloud/nextcloudpi/commit/f047cf7) (2018-05-22) fix php cli tmpdir for running instances
+[v0.56.9](https://github.com/nextcloud/nextcloudpi/commit/72cd80d) (2018-05-24) remove old systemd timer config in running systems
+
+[v0.56.8](https://github.com/nextcloud/nextcloudpi/commit/cf42bf8) (2018-05-22) fix php cli tmpdir for running instances
 
 [v0.56.7 ](https://github.com/nextcloud/nextcloudpi/commit/54da658) (2018-05-22) move NC httpd logs to /var/log